confirm delete operation in table components

+++ src/app/catalog/catalog-table.component.ts
@@ -15,6 +15,7 @@
 
 import { Component, EventEmitter, OnInit, Output } from '@angular/core';
 import { firstValueFrom, Observable, ReplaySubject } from 'rxjs';
+import { ConfirmDialogService } from '../shared/components/confirm-dialog.component';
 import { ITableColumn } from '../shared/components/table.component';
 import { Catalog, CatalogService } from '../shared/services/catalog.service';
 import { CatalogDialogService } from './catalog-dialog.component';
@@ -39,7 +40,8 @@ export class CatalogTableComponent implements OnInit {
 
   constructor(
     protected _catalogService: CatalogService,
-    protected _catalogDialogService: CatalogDialogService
+    protected _catalogDialogService: CatalogDialogService,
+    protected _confirmDialogService: ConfirmDialogService
   ) {}
 
   async ngOnInit(): Promise<void> {
@@ -63,8 +65,15 @@ export class CatalogTableComponent implements OnInit {
   }
 
   async onDeleteCatalog(catalog: Catalog): Promise<void> {
-    await firstValueFrom(this._catalogService.deleteCatalog(catalog.id));
-    await this.onReloadCatalogs();
+    const confirmDialogRef = this._confirmDialogService.openConfirmDialog(
+      'Delete Catalog',
+      `Do you really want to delete the catalog "${catalog.title}"?`
+    );
+    const confirmed = await firstValueFrom(confirmDialogRef.afterClosed());
+    if (confirmed) {
+      await firstValueFrom(this._catalogService.deleteCatalog(catalog.id));
+      await this.onReloadCatalogs();
+    }
   }
 
   async onReloadCatalogs(): Promise<void> {
